refactor: simplify usage of app and local app configs
