+++ config/index.js
@@ -1,0 +1,6 @@
+const defaultConfig = require('./main');
+const localConfig = require('./main.local');
+module.exports.config = {
+  ...defaultConfig,
+  ...localConfig
+};
