+++ config/main.js
@@ -1,6 +1,6 @@
 /** General Configurations Like PORT, HOST names and etc... */
 
-var config = {
+const config = {
   env: process.env.NODE_ENV || 'development',
   host: process.env.HOST || 'localhost',
   port: process.env.PORT || 8889,
