+++ src/app/redux/configureStore.ts
@@ -4,12 +4,10 @@ import {createLogger} from "redux-logger";
 import {router5Middleware} from "redux-router5";
 import createSagaMiddleware, {END, Task} from "redux-saga";
 import {Router} from "router5";
+import {config as appConfig} from "../../../config";
 import {IStore} from "./IStore";
 import {sentryMiddleware} from "./middlewares/sentryMiddleware";
 import rootReducer from "./rootReducer";
-
-const appConfig = require("../../../config/main");
-const localConfig = require("../../../config/main.local");
 
 interface IExtendedStore extends Store<Partial<IStore>> {
   runSaga: (rootSaga: any) => Task;
@@ -17,8 +15,6 @@ interface IExtendedStore extends Store<Partial<IStore>> {
 }
 
 export function configureStore(router: Router, initialState?: Partial<IStore>): IExtendedStore {
-
-  const mergedConfig = {...appConfig, ...localConfig};
   const sagaMiddleware = createSagaMiddleware();
   const middlewares: Middleware[] = [
     router5Middleware(router),
@@ -26,21 +22,21 @@ export function configureStore(router: Router, initialState?: Partial<IStore>): 
   ];
 
   /** Add Only Dev. Middlewares */
-  if (mergedConfig.env !== "production" && process.env.BROWSER) {
+  if (appConfig.env !== "production" && process.env.BROWSER) {
     const logger = createLogger();
     middlewares.push(logger);
   }
 
-  if (mergedConfig.sentry && mergedConfig.sentry.dsn && process.env.BROWSER) {
+  if (appConfig.sentry && appConfig.sentry.dsn && process.env.BROWSER) {
     middlewares.unshift(sentryMiddleware);
 
     Sentry.init({
-      dsn: mergedConfig.sentry.dsn,
-      ...mergedConfig.sentry.options
+      dsn: appConfig.sentry.dsn,
+      ...appConfig.sentry.options
     });
   }
 
-  const composeEnhancers = (mergedConfig.env !== "production" &&
+  const composeEnhancers = (appConfig.env !== "production" &&
     typeof window === "object" &&
     (typeof window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ === "function") &&
     window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__({shouldHotReload: false})) || compose;
@@ -49,7 +45,7 @@ export function configureStore(router: Router, initialState?: Partial<IStore>): 
     applyMiddleware(...middlewares)
   ));
 
-  if (mergedConfig.env === "development" && (module as any).hot) {
+  if (appConfig.env === "development" && (module as any).hot) {
     (module as any).hot.accept("./rootReducer", () => {
       store.replaceReducer((require("./rootReducer").default));
     });
